More work on #7: Handle losing the game
After the user steps on a mine:
All the non-marked mine cells are revealed through the character * and are colored in red. The cell of the mine that has been stepped on is blinking.
The character - is displayed on the non-mine marked cells. Their color remains orange/yellow instead of red, to avoid confusion (the color of the mines is red as well) – searching for a better solution.
A text is displayed on the right side of the board, below the counter #3, in red: "YOU LOST".