Fixed: Undesired cells are automatically stepped on
When the user presses on a stepped cell located at the edge of the board, two additional cells may be stepped on:
(x+2; y) and (x-2; y) - if the cell is located on the top/bottom of the board;
(x; y+2) and (x; y-2) - if the cell is located on the right/bottom of the board.

Cause
The cells outside the physical board are declared as unmarked, unstepped and have the value 0 (neighbour-mines) assigned to them. Because in pressed-a-stepped-on-cell it is not verified whether the cell is valid (on the physical board) or not, the 3 cells outside the board are pressed. If one of those cells has the value 0, the function  ZERO_PRESSED is called. Thus, all its adjacent cells are pressed, including the undesired ones. The process does not repeat because, after the first zero is pressed, the function is called again only if the cell is valid.

Fix
Verify the validity of the cell before stepping on it in pressed-a-stepped-on-cell.