Fixed: Some cells display a value greater than expected
In the 3rd level of difficulty (maximum rows, maximum columns), some cells located on the left/right edge of the board display a value greater than the actual number of mine neighbours they have.

Cause & Mechanism
When an operation is performed on a cell located outside the physical board, the exact same operation is performed on the last cell of the previous row or the first cell of the next row – (-1; y) ⇒ (xmax-1; y-1);  (xmax; y) ⇒ (0; y+1). (WHY?)
Therefore, a mine placed on the first/last column of the board determines an increase in the displayed value of mine neighbours of 1, 2 or 3 cells located on the last/first column.

Fix
In the process of generating the board, verify the validity of the cell before increasing its value.