Work on #3 and #7: Not enough mines are laid on board – Mac OS & Linux
Brief description of the problem
The number of mines laid on the board is smaller than the expected number of total mines. As the end of the game was defined using the total number of mines, problems emerged on #7. Moreover, the value displayed on the counter of marked mines #3 represents the expected number of mines, not the actual number of laid mines. These issues were observed on Linux and Mac OS, but were absent on Windows.

Cause
The root of the problem is in the condition evaluated in order to decide whether a mine will be placed in a particular cell or not. The value of RAND_MAX (the maximum value of the pseudo-random numbers returned by rand()) is library-dependent, but it is guaranteed to be at least 32767. The program was originally tested for this value. No problem here. On the other hand, on Ubuntu 20.04, RAND_MAX is 2147483647 (the maximum value of an int). After multiplying it by the number of mines waiting to be placed on the board, an integer overflow occurred (both RAND_MAX and the number of mines are ints), messing up the probability.

Solution
The new formula is rearranged so that this situation is avoided while preserving its mathematical meaning.
The right-hand side and the left-hand side of the expression have to be converted into rational numbers because, otherwise, they can only take the value 0 or 1 (ridiculously rare, the RHS can be greater than 1). Float could not be used in this type of conversion because it has 23 bits of significand (it represents numbers in the range [-2^23, 2^23]), while int has 31 (it represents numbers in the range [-2^31, 2^31-1]). Double works fine, with 52 bits of significand.