Delete unnecessary condition evaluated before the probability of mines
Brief description
There is no need to verify that total_mines - laid_mines is different from 0. It was intended to be a solution to "Process returned -1073741676 (0xC0000094)" (division by zero). However, it only hid the problem.

Cause
The former formula used for calculating the probability of a mine being in a certain cell is incorrect. It takes into account only the cells located after its current position. An attempt to divide by zero occurs when the expression is evaluated for the last cell of the board. It has zero cells after it therefore the denominator is 0. Most of the time, probabilistically speaking, all the mines have been placed before the last cell. By making sure that there still are mines to be placed, most of the time the expression is not evaluated for the last cell.

Solution
The correct formula for the probability takes into account the cell it is calculating the probability for (thus the new denominator is one greater than the old denominator). In the extreme case, when the expression is evaluated for the last cell of the board, the denominator is 1, as opposed to 0. Division by zero is not attempted.